backend/requirements.txt
absl-py==0.8.1
aioredis==1.2.0
api2pdf==0.0.11
asgiref==3.2.2
asn1crypto==0.24.0
astor==0.8.0
async-timeout==3.0.1
attrs==19.1.0
authy==2.2.6
autobahn==19.9.2
Automat==0.7.0
Babel==2.8.0
beautifulsoup4==4.8.0
bleach==1.5.0
boto==2.49.0
boto3==1.13.7
botocore==1.16.7
bs4==0.0.1
cachetools==3.1.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.9.11
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
cryptography==2.7
cssselect2==0.2.2
cycler==0.10.0
daphne==2.3.0
dash==1.3.0
dash-core-components==1.2.0
dash-html-components==1.0.1
dash-renderer==1.1.0
dash-table==4.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.4
django-autocomplete-light==3.4.1
django-autoslug-iplweb==1.9.4
django-dash==0.5.6
django-decorator-include==3.0
django-easy-pdf==0.1.1
django-extensions==2.2.9
django-formtools==2.2
django-haystack==2.8.1
django-heroku==0.3.1
django-nine==0.2.2
django-otp==0.9.0
django-otp-yubikey==0.5.2
django-pandas==0.6.1
django-phonenumber-field==3.0.1
django-picklefield==2.0
django-plotly-dash==1.0.1
django-redis==4.10.0
django-storages==1.9.1
django-tables2==2.1.0
django-tinymce==2.8.0
django-toolbelt==0.0.1
django-twilio==0.13.0.post1
django-two-factor-auth==1.11.0
django-weasyprint==0.5.4
django-wkhtmltopdf==3.2.0
djangorestframework==3.10.3
docutils==0.15.2
dpd-components==0.1.0
easy-thumbnails==2.6
elasticsearch==7.0.5
et-xmlfile==1.0.1
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.1
Flask-Compress==1.4.0
fpdf==1.7.2
future==0.17.1
gast==0.2.2
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
graphviz==0.14
grpcio==1.24.3
gunicorn==19.9.0
h5py==2.10.0
hiredis==1.0.0
html5lib==0.9999999
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jmespath==0.9.5
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==7.2.0
msgpack==0.6.1
numpy==1.16.4
oauthlib==3.1.0
openpyxl==2.6.3
opt-einsum==3.1.0
ordereddict==1.1
packaging==19.1
pandas==0.25.0
pdfkit==0.6.1
phonenumbers==8.12.2
pif==0.8.2
Pillow==6.1.0
placeholder==1.0
plotly==4.1.1
pluggy==0.13.0
protobuf==3.10.0
psutil==5.7.0
psycopg2==2.8.3
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.9.7
pydotplus==2.0.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pyparsing==2.4.2
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.0
python-pdf==0.37
pytz==2019.2
PyYAML==5.1.2
qrcode==6.1
redis==3.3.8
reportlab==3.5.26
requests==2.22.0
requests-oauthlib==1.2.0
retrying==1.3.3
rsa==4.0
s3transfer==0.3.3
scipy==1.3.1
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
static3==0.7.0
termcolor==1.1.0
tinycss2==1.0.2
toml==0.10.0
tornado==6.0.3
tox==3.14.0
transliterate==1.10.2
twilio==6.44.1
Twisted==19.7.0
txaio==18.8.1
Unidecode==1.1.1
urllib3==1.25.3
virtualenv==16.7.5
vishap==0.1.5
WeasyPrint==50
webencodings==0.5.1
Werkzeug==0.15.6
whitenoise==4.1.3
wkhtmltopdf-pack-ng==0.12.3.0
wrapt==1.11.2
xhtml2pdf==0.2.3
xlrd==1.2.0
YubiOTP==0.2.2.post1
zipp==0.6.0
zope.interface==4.6.0
